[Feature #165412925] user should follow each other
