+++ src/controllers/AuthLocalController.js
@@ -131,19 +131,6 @@ export default class AuthLocalController {
   }
 
   /**
-   * @description function to activate user account
-   * @param {object} req
-   * @param {object} res
-   * @returns {object} it return true if account verified otherwise it return false
-   */
-  static async reset(req, res) {
-    return res.status(status.OK).json({
-      message: 'Your request accepted use the below token to rest your password',
-      token: req.params.token
-    });
-  }
-
-  /**
    * @param  {object} req
    * @param  {object} res
    * @return {object} return an object containing the confirmation message
@@ -193,102 +180,5 @@ export default class AuthLocalController {
         .status(status.OK)
         .json({ isUpdated, message: 'Success! your password has been changed.' })
       : res.status(status.NOT_MODIFIED).json({ errors: 'Password not updated' });
-  }
-
-  // follow user
-  /**
-   * @description function to create user follows
-   * @param {object} req request from user
-   * @param {object} res server response
-   * @returns {object} true
-   */
-  static async follow(req, res) {
-    const { username } = req.params;
-    const checkUser = await User.findOne({ username });
-    if (checkUser.id === req.user.id) {
-      return res
-        .status(status.BAD_REQUEST)
-        .json({ errors: { follow: 'You can not follow your self ' } });
-    }
-    const follow = await User.follow.add({
-      followed: checkUser.id,
-      userId: req.user.id
-    });
-    if (follow.errors) {
-      return follow.errors.name === 'SequelizeUniqueConstraintError'
-        ? res
-          .status(status.EXIST)
-          .send({ errors: { follow: `You are already following "${username}"` } })
-        : res.status(status.SERVER_ERROR).json({ errors: 'oops, something went wrong' });
-    }
-    return res.status(status.CREATED).json({
-      message: `now you are following ${checkUser.username}`
-    });
-  }
-
-  // unFollow user
-  /**
-   * @description function to allow user to unfollow users
-   * @param {object} req user request
-   * @param {object} res response from server
-   * @returns {object} true
-   */
-  static async unfollow(req, res) {
-    const [username, user] = [req.params.username, req.user];
-    const checkUser = await User.findOne({ username });
-
-    const hasUnfollowed = Object.keys(checkUser).length
-      ? await User.follow.remove({ userId: user.id, followed: checkUser.id })
-      : null;
-
-    if (hasUnfollowed && hasUnfollowed.errors) {
-      return res.status(status.SERVER_ERROR).json({ errors: 'oops, something went wrong!!' });
-    }
-    return hasUnfollowed
-      ? res.status(status.OK).json({
-        message: `you unfollowed ${username}`
-      })
-      : res
-        .status(status.BAD_REQUEST)
-        .json({ errors: { follow: `you are not following "${username}"` } });
-  }
-
-  /**
-   * @description function to fetch users'followers
-   * @param {object} req
-   * @param {object} res
-   * @returns {object} followers
-   */
-  static async followers(req, res) {
-    const { id } = req.user;
-    const followers = await User.follow.getAll({ followed: id });
-    return followers.length
-      ? res.status(status.OK).json({
-        message: 'Followers',
-        followers: followers.map(follower => delete follower.get().followedUser && follower)
-      })
-      : res.status(status.NOT_FOUND).json({
-        errors: { follows: 'You do not have followers for know' }
-      });
-  }
-
-  /**
-   * @description function to fetch all authors who user follow
-   * @param {object} req
-   * @param {object} res
-   * @returns {object} followers
-   */
-  static async following(req, res) {
-    const following = await User.follow.getAll({ userId: req.user.id });
-    const follows = following.map(followed => delete followed.get().follower && followed);
-    if (following.length) {
-      return res.status(status.OK).json({
-        message: 'Following',
-        following: follows
-      });
-    }
-    return res.status(status.NOT_FOUND).json({
-      errors: { follows: 'You do not follow any one' }
-    });
   }
 }
