+++ src/middlewares/validateLogin.js
@@ -16,9 +16,11 @@ export default async (req, res, next) => {
 
   Object.keys(errors).forEach(key => errors[key] || delete errors[key]);
 
-  return !Object.keys(errors).length
-    ? next()
-    : res.status(status.BAD_REQUEST).json({
-      errors
-    });
+  return (
+    (Object.keys(errors).length
+      && res.status(status.BAD_REQUEST).json({
+        errors
+      }))
+    || next()
+  );
 };
