+++ templates/html/admin.html
@@ -18,6 +18,7 @@
           Admin
         </h1>
         <div class="large-padding">
+          <div class="bold loggedInUser"></div>
           <div class="radius-2 medium-padding card">
             <input
               type="url"
