+++ templates/html/follow.html
@@ -18,6 +18,7 @@
           Follow
         </h1>
         <div class="large-padding">
+          <div class="bold loggedInUser"></div>
           <div class="radius-2 medium-padding card">
             <input
               type="url"
@@ -178,7 +179,6 @@
 
           switch (verb) {
             case 'PATCH':
-              console.log(URL);
               await follow(URL);
               break;
             case 'POST':
